Use doubles for the shape api

+++ plugin/src/main/java/com/quartercode/quarterbukkit/api/shape/Cuboid.java
@@ -202,7 +202,7 @@ public class Cuboid implements Shape {
     }
 
     @Override
-    public Collection<Vector> getContent(float distance) {
+    public Collection<Vector> getContent(double distance) {
 
         Collection<Vector> vectors = new ArrayList<Vector>();
